Queue parameters set via begin()
- Parameters are sized using malloc to allow for sizing via the API.

- IPqueue is now 4 byte array instead of IPAddress because I couldn't get
IPAddress array to work with malloc.  This requires conversion of the
IPAddress parameter of queue() to the byte array and conversion back to
IPAddress before the EtherEvent.send() but the byte array is 4 bytes and
the IPAddress is 6 byte so there is a small decrease in the memory taken
up by the queue.

- getNode() is now a template(and so had to be moved to EtherEventQueue.h)
to allow use of the same function definition for IPAddress and byte
array type IP addresses.

- EtherEvent buffer sizes are now automatically sized to match the
EtherEventQueue received event buffer sizes.