Don't send acks automatically
It is now the user's responsibility to send acks. I made this change
because it was inefficient to send acks for all events(other that
keepalive and ack) when the event might not require an ack. When a send
is successful that is a preliminary ack and then if a final ack is
needed then that should come from the user's code to be sure that the
event has been processed at that level.