Disable MD5.h include for ESP8266
The ESP8266 core contains a file named md5.h. On filename case
insensitive operating systems this file is included instead of MD5.h,
which causes compilation to fail. I use the MD5Builder library included
with the ESP8266 core instead of ArduinoMD5 so the include of MD5.h is
not necessary.