Move test sketches, etc. to the extras folder
The library specification says "The content of the extras folder is totally ignored by the IDE". Unfortunately this is wrong. Any sketch in the library folder will appear on the Tools > Examples menu. However, it is best to follow the specificaation in hopes that eventually the IDE will follow suit.